960-expression-service/src/main/java/io/harness/expression/configuration/util/ExpressionServiceConfigurationUtils.java
/*
 * Copyright 2020 Harness Inc. All rights reserved.
 * Use of this source code is governed by the PolyForm Shield 1.0.0 license
 * that can be found in the licenses directory at the root of this repository, also available at
 * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
 */

package io.harness.expression.configuration.util;

import io.harness.expression.app.ExpressionServiceConfiguration;

import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
import java.io.IOException;
import java.io.InputStream;
import lombok.experimental.UtilityClass;

@UtilityClass
public class ExpressionServiceConfigurationUtils {
  public ExpressionServiceConfiguration getApplicationConfiguration(InputStream config) throws IOException {
    ObjectMapper objectMapper = new ObjectMapper(new YAMLFactory());
    return objectMapper.readValue(config, ExpressionServiceConfiguration.class);
  }
}
